+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - use Variable and Labels in "match()" step
 - use Variable in "return()" step
+- return a Cypher statement without return clause
 
 ## [0.0.2]  2022-07-20
 ### Added